Imported Row component and updated imports
