+++ src/ui/index.js
@@ -2,7 +2,6 @@
 import Accordion from './Accordion'
 import BackgroundContentBlock from './BackgroundContentBlock'
 import Block from './Block'
-import BlockLayout from './BlockLayout'
 import Button from './Button'
 import ButtonRow from './ButtonRow'
 import Carousel from './Carousel'
@@ -12,14 +11,13 @@ import GroupBlock from './GroupBlock'
 import Layout from './Layout'
 import Loader from './Loader'
 import Media from './Media'
-// import Row from './Row'
+import Row from './Row'
 
 
 export { 
     Accordion, 
     BackgroundContentBlock, 
     Block, 
-    BlockLayout, 
     Button, 
     ButtonRow, 
     Carousel,
@@ -28,7 +26,8 @@ export {
     GroupBlock,
     Layout,
     Loader,
-    Media
+    Media,
+    Row
 }
 
 export {
